+++ src/sampling/error_dither.py
